Feature/calling improvements (#15)
* fixed walter migrations to not run when the app is in production mode

* pulled in libphonenumber, using it to validate phone numbers in call model before publishing them to kafka. Still need to alter the CallReader to pull the international field from calls table in stats

* adjusted migrations, tests, and the call model and reader to use the international column on the stats database calls table

* working on parsing methods on Call model, going to write a test for it. going to rebase this branch on top of develop first to get the latest kafka producer and consumer changes

* created call test for publishing to kafka, that method also validates phone numbers. Rewrote call factory a bit

* added a couple more tests

* fixed typo in migrations and made conditional simpler per ethans suggestion

* not filtering calls marked valid=false in stats. It's my understanding that forbes isnt concerned with what the old stats considered a valid call. He wants those calls anyway

* edited call test to match change

* moved the phone utility to an accessor. Wrote new test to make sure international numbers were being parsed correctly

+++ database/migrations/2019_08_23_191505_create_SendOut_table.php
@@ -1,5 +1,6 @@
 <?php
 
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -13,15 +14,17 @@ class CreateSendOutTable extends Migration
      */
     public function up()
     {
-        Schema::connection('sqlite_walter_test')->dropIfExists('SendOut');
-        Schema::connection('sqlite_walter_test')->create('SendOut', function (Blueprint $table) {
-            $table->bigIncrements('soid');
-            $table->dateTime('DateSent');
-            $table->integer('Consultant');
-            $table->boolean('firstResume');
+        if (!App::environment('production')) {
+            Schema::connection('sqlite_walter_test')->dropIfExists('SendOut');
+            Schema::connection('sqlite_walter_test')->create('SendOut', function (Blueprint $table) {
+                $table->bigIncrements('soid');
+                $table->dateTime('DateSent');
+                $table->integer('Consultant');
+                $table->boolean('firstResume');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -31,6 +34,8 @@ class CreateSendOutTable extends Migration
      */
     public function down()
     {
-        Schema::connection('sqlite_walter_test')->dropIfExists('SendOut');
+        if (!App::environment('production')) {
+            Schema::connection('sqlite_walter_test')->dropIfExists('SendOut');
+        }
     }
 }
