+++ database/migrations/2019_08_27_192951_create_job_order_interview_table.php
@@ -1,5 +1,6 @@
 <?php
 
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -13,14 +14,16 @@ class CreateJobOrderInterviewTable extends Migration
      */
     public function up()
     {
-        Schema::connection('sqlite_walter_test')->dropIfExists('jobOrder_interview');
-        Schema::connection('sqlite_walter_test')->create('jobOrder_interview', function (Blueprint $table) {
-            $table->bigIncrements('intID');
-            $table->dateTime('dateCreated');
-            $table->integer('consultant');
+        if (!App::environment('production')) {
+            Schema::connection('sqlite_walter_test')->dropIfExists('jobOrder_interview');
+            Schema::connection('sqlite_walter_test')->create('jobOrder_interview', function (Blueprint $table) {
+                $table->bigIncrements('intID');
+                $table->dateTime('dateCreated');
+                $table->integer('consultant');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -30,6 +33,8 @@ class CreateJobOrderInterviewTable extends Migration
      */
     public function down()
     {
-        Schema::connection('sqlite_walter_test')->dropIfExists('jobOrder_interview');
+        if (!App::environment('production')) {
+            Schema::connection('sqlite_walter_test')->dropIfExists('jobOrder_interview');
+        }
     }
 }
