+++ database/migrations/2019_08_28_153441_create_stats_calls_table.php
@@ -1,5 +1,6 @@
 <?php
 
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -13,20 +14,23 @@ class CreateStatsCallsTable extends Migration
      */
     public function up()
     {
-        Schema::connection('sqlite_stats_test')->dropIfExists('calls');
-        Schema::connection('sqlite_stats_test')->create('calls', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('user_id')->unsigned();
-            $table->boolean('valid');
-            $table->bigInteger('dialed_number')->unsigned();
-            $table->enum('type', ['Incoming', 'Outgoing', 'Transfer'])->nullable();
-            $table->dateTime('date');
-            $table->integer('duration')->unsigned();
-            $table->string('raw')->nullable()->default(null);
+        if (!App::environment('production')) {
+            Schema::connection('sqlite_stats_test')->dropIfExists('calls');
+            Schema::connection('sqlite_stats_test')->create('calls', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->bigInteger('user_id')->unsigned();
+                $table->boolean('valid');
+                $table->bigInteger('dialed_number')->unsigned();
+                $table->boolean('international');
+                $table->enum('type', ['Incoming', 'Outgoing', 'Transfer'])->nullable();
+                $table->dateTime('date');
+                $table->integer('duration')->unsigned();
+                $table->string('raw')->nullable()->default(null);
 
-            $table->timestamps();
-            $table->softDeletes();
-        });
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
@@ -36,6 +40,8 @@ class CreateStatsCallsTable extends Migration
      */
     public function down()
     {
-        Schema::connection('sqlite_stats_test')->dropIfExists('calls');
+        if (!App::environment('production')) {
+            Schema::connection('sqlite_stats_test')->dropIfExists('calls');
+        }
     }
 }
